feat/#44: 게시물 상세조회 API 댓글 연동

+++ app/src/main/java/com/konkuk/arabyte_aos/data/dataremote/datasource/NoticeBoardRemoteDataSource.kt
@@ -13,6 +13,6 @@ interface NoticeBoardRemoteDataSource {
     ): Response<GetNoticeBoardListResponseDto>
 
     suspend fun getNoticeBoardDetail(
-        articleId: Int,
+        articleId: Long,
     ): Response<GetNoticeBoardDetailResponseDto>
 }
