+++ app/src/main/java/com/konkuk/arabyte_aos/data/dataremote/datasourceimpl/NoticeBoardRemoteDataSourceImpl.kt
@@ -25,7 +25,7 @@ class NoticeBoardRemoteDataSourceImpl
                 sort = sort,
             )
 
-        override suspend fun getNoticeBoardDetail(articleId: Int): Response<GetNoticeBoardDetailResponseDto> =
+        override suspend fun getNoticeBoardDetail(articleId: Long): Response<GetNoticeBoardDetailResponseDto> =
             service.getNoticeBoardDetail(
                 articleId = articleId,
             )
