+++ app/src/main/java/com/konkuk/arabyte_aos/data/dataremote/service/NoticeBoardService.kt
@@ -19,6 +19,6 @@ interface NoticeBoardService {
 
     @GET("/$ARTICLES/{articleId}")
     suspend fun getNoticeBoardDetail(
-        @Path("articleId") articleId: Int,
+        @Path("articleId") articleId: Long,
     ): Response<GetNoticeBoardDetailResponseDto>
 }
