+++ app/src/main/java/com/konkuk/arabyte_aos/domain/repository/NoticeBoardRepository.kt
@@ -12,6 +12,6 @@ interface NoticeBoardRepository {
     ): Result<NoticeBoardList>
 
     suspend fun getNoticeBoardDetail(
-        articleId: Int,
+        articleId: Long,
     ): Result<NoticeBoardDetail>
 }
