+++ app/src/main/java/com/konkuk/arabyte_aos/presentation/ui/noticeboarddetail/NoticeBoardDetailViewModel.kt
@@ -25,7 +25,7 @@ class NoticeBoardDetailViewModel
 
         private fun getNoticeBoardDetail(articleId: Long) {
             viewModelScope.launch {
-                getNoticeBoardDetailUseCase(articleId.toInt())
+                getNoticeBoardDetailUseCase(articleId)
                     .onSuccess { noticeBoardDetail ->
                         setState {
                             copy(
